Renaming "cache" to "cached" for clarity
It is misleading to call this `cache` as it is loading the `grunt-cached` plugin.  All of the other variable names accurately represent the module that's being loaded.  Additionally, there is a `grunt-cache` plugin.  If I were to assume that people get things right most of the time, I'd try to first write my code to load `grunt-cache` because that variable is right twice and I would guess that the typo was just in the module name.