Set chokidar option ignoreIntial: true by default
Chokidar by default emits `add` events as it initially scans the file tree it is asked to watch. This can be suppressed by setting the `ignoreInitial: true` option, which is desirable here to avoid kicking off the task for each matched file upon startup.

Users retain the ability to specify `ignoreInitial: false` if that is the behavior they want.