Docs: Add is it maintained badges
Is it maintained? is an open source project whose goal is to monitor
open source projects, certifying how well it is maintained.

The metrics are:

  - RESOLUTION TIME: Median time needed to close an issue
  - OPEN PERCENTAGE: Percentage of open issues.

Some issues are ignored based on their labels (example of ignored
labels: "enhancement", "feature"…). For the median time needed to close
an issue, issues older than 6 months are ignored too.

References:

  - IS IT MAINTAINED? Check the activity of open source projects
    https://isitmaintained.com/#about